Add files via upload
This is just a component build, there is no content. A simple page that has a h1 nav. Using javaScript in html to make nav disappear on scroll-down, and appear with a shadow on scroll-up.
This might be useful (on blog sites, for example) where the reader doesn't have to scroll back to 0 to access the nav, or have the fixed nav block viewport while reading.